+++ src/main/java/techcourse/myblog/domain/exception/CommentUpdateFailedException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.exception;
 
 public class CommentUpdateFailedException extends RuntimeException {
     public static final String COMMENT_UPDATE_FAIL_ERROR_MSG = "댓글을 수정할 수 없습니다";
