+++ src/main/java/techcourse/myblog/domain/exception/InvalidCommentException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.exception;
 
 public class InvalidCommentException extends RuntimeException {
     public static final String INVALID_COMMENT_ERROR_MSG = "댓글은 비어있을 수 없습니다";
