+++ src/main/java/techcourse/myblog/domain/exception/InvalidUserException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.domain;
+package techcourse.myblog.domain.exception;
 
 public class InvalidUserException extends RuntimeException {
     public InvalidUserException(String msg) {
