+++ src/main/java/techcourse/myblog/service/UserWriteService.java
@@ -5,6 +5,7 @@ import org.springframework.transaction.annotation.Transactional;
 import techcourse.myblog.domain.User;
 import techcourse.myblog.domain.repository.UserRepository;
 import techcourse.myblog.dto.UserDto;
+import techcourse.myblog.service.exception.SignUpFailedException;
 
 @Service
 @Transactional
