+++ src/main/java/techcourse/myblog/service/exception/MismatchAuthorException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.exception;
 
 public class MismatchAuthorException extends RuntimeException {
     public static final String MISMATCH_AUTHOR_ERROR_MSG = "작성자가 아닙니다";
