+++ src/main/java/techcourse/myblog/service/exception/NotFoundArticleException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.exception;
 
 public class NotFoundArticleException extends RuntimeException {
     public static final String NOT_FOUND_ARTICLE_ERROR_MSG = "존재하지 않는 게시글입니다";
