+++ src/main/java/techcourse/myblog/service/exception/NotFoundCommentException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.exception;
 
 public class NotFoundCommentException extends RuntimeException {
     public static final String NOT_FOUND_COMMENT_ERROR_MSG = "존재하지 않는 댓글입니다";
