+++ src/main/java/techcourse/myblog/service/exception/SignUpFailedException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.service;
+package techcourse.myblog.service.exception;
 
 public class SignUpFailedException extends RuntimeException {
     public SignUpFailedException(String msg) {
