+++ src/main/java/techcourse/myblog/web/exception/InvalidUserSessionException.java
@@ -1,9 +1,9 @@
-package techcourse.myblog.web.support;
+package techcourse.myblog.web.exception;
 
 public class InvalidUserSessionException extends RuntimeException {
     private static final String INVALID_USER_SESSION_ERROR_MSG = "로그인 세션이 유효하지 않습니다";
 
-    InvalidUserSessionException() {
+    public InvalidUserSessionException() {
         super(INVALID_USER_SESSION_ERROR_MSG);
     }
 }
