+++ src/main/java/techcourse/myblog/web/exception/UnauthorizedRequestException.java
@@ -1,4 +1,4 @@
-package techcourse.myblog.web.controller;
+package techcourse.myblog.web.exception;
 
 public class UnauthorizedRequestException extends RuntimeException {
     public static final String UNAUTHORIZED_REQUEST_ERROR_MSG = "권한이 없는 요청입니다";
