+++ src/test/java/techcourse/myblog/domain/ArticleTest.java
@@ -1,7 +1,7 @@
 package techcourse.myblog.domain;
 
 import org.junit.jupiter.api.Test;
-import techcourse.myblog.service.MismatchAuthorException;
+import techcourse.myblog.service.exception.MismatchAuthorException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
