+++ src/test/java/techcourse/myblog/domain/CommentTest.java
@@ -1,6 +1,8 @@
 package techcourse.myblog.domain;
 
 import org.junit.jupiter.api.Test;
+import techcourse.myblog.domain.exception.CommentUpdateFailedException;
+import techcourse.myblog.domain.exception.InvalidCommentException;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
