+++ src/test/java/techcourse/myblog/domain/UserTest.java
@@ -3,6 +3,7 @@ package techcourse.myblog.domain;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
+import techcourse.myblog.domain.exception.InvalidUserException;
 
 import java.util.stream.Stream;
